requirements.txt
Django==4.2
djangorestframework==3.14.1
django-extensions==3.2.3
django-phone-field==1.8.1
django-phonenumber-field==7.0.0
Jinja2==3.1.2
